Provide for annotations to kubernetes service
This change allows the user to provide annotations to the k8s svc for
drone.

In our case, we rely on service annotations to interact with GCP NEGs
via a controller that reads them.

This adds to the templates/service.yaml an optional 'annotations' field
that can be populated by the user, or left empty by default.

This shouldn't be conflaited with the ingress annotations. We
specifically do not use k8s ingress and we should be able to articulate
annotations for either, both, or none of the service/ingress resources.