Allow configuration of deployment replica count
Drone server is a singleton, but it may be desirable to deploy the helm chart with
 a replica count of 0. A scenario here would be deploying a cold-standby copy of
 drone to a different cluster/region/namespace/etc

In our case, we intend to keep a cold-standby deployed but spun down,
and the desire is to flip the replica count to 1 in the failover site
and flip to 0 in the production site after the drone DB has been cloned
to the failover region.

Since the service is a singleton, it's important to ensure this value
can only ever be 1 or 0. Helm templating ensures the following:

Numbers greater than or equal to 1 will be set to 1
Numbers less than 1 will be set to 0